CommissionMembre: Fixed version_relation models+ids retrieval (relates #118)

+++ sql/001_versions.sql
@@ -30,7 +30,7 @@ CREATE TABLE versions_data (
 
 DROP TABLE IF EXISTS versions_relations;
 CREATE TABLE versions_relations (
-    id MEDIUMINT UNSIGNED NOT NULL AUTO_INCREMENT,
+    id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
     version_id INT UNSIGNED NOT NULL,
     table_name VARCHAR(255) NOT NULL,
     model_name VARCHAR(255) NOT NULL,
